demos: fix compilation warnings (#3867)

+++ demos/text_detection_demo/cpp/include/cnn.hpp
@@ -32,7 +32,7 @@ protected:
     ov::Core& m_core;
     cv::Size m_new_input_resolution;
     bool use_auto_resize;
-    int m_channels;
+    size_t m_channels;
     cv::Size m_input_size;
     std::string m_input_name;
     std::vector<std::string> m_output_names;
